==> notebooks/3-Split-Data.ipynb <==
# Separação dos Dados

import pandas as pd
from sklearn.model_selection import train_test_split

# Dados

df = pd.read_csv('../data/raw/fraud_dataset_v2.csv', index_col=0)

df.head()

# Separação

Para treino será utilizado 80% do período [20200308, 20200407].  
Para validação será utilizado 20% restante do período de treino [20200308, 20200407].  
Para teste será utilizado 100% do período [20200408, 20200421].  

Vou utilizar uma técnica de out-of-time para o período de teste do modelo e uma técnica de out-of-sample para as partes de hipertunning/substituição de nulos/feature selection, evitando data leak.  

df_train = df[
    (df['fecha'].astype('datetime64[s]') >= pd.to_datetime('2020-03-08')) &
    (df['fecha'].astype('datetime64[s]') <= pd.to_datetime('2020-04-07'))             
]

df_train, df_valid = train_test_split(df_train, test_size=0.2, stratify=df_train['fraude'], random_state=777)

df_test = df[
    (df['fecha'].astype('datetime64[s]') >= pd.to_datetime('2020-04-08'))         
].reset_index(drop=True)

df_train.to_parquet(
    '../data/raw/fraud_dataset_v2_train.parquet.gzip',
    compression='gzip',
    index=False
)

df_valid.to_parquet(
    '../data/raw/fraud_dataset_v2_valid.parquet.gzip',
    compression='gzip',
    index=False
)

df_test.to_parquet(
    '../data/raw/fraud_dataset_v2_test.parquet.gzip',
    compression='gzip',
    index=False
)
