==> notebooks/4-Data-Basic-Process.ipynb <==
# Preparação dos Dados

import pandas as pd
import numpy as np
import sys
import yaml
sys.path.append('../src/')

from utils.transformers import FixFeaturesType, FixFeaturesMissing

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 100)

# Data

df = pd.read_csv('../data/raw/fraud_dataset_v2.csv', index_col=0)

df.head()

df_info = df.describe(include='all').T
df_info['tipo_coluna'] = df.dtypes
df_info['qtd_unicos'] = df.nunique()
df_info['tipo_final'] = df_info['qtd_unicos'].apply(lambda x: 'categorica' if x <= 70 else 'numerica')
df_info

Avaliando as tabelas acima será necessário vamos realizar alguns ajustes de tipo e nulo.  
Ajustes de Tipo:  
- Ajustar coluna de data 'fecha' para tipo timestamp  

Ajuste de Nulos:  
- Ajustar colunas ['d', 'g', 'o', 'q'], para as colunas categóricas vou criar uma nova categoria indicando o nulo. 
- Ajustar colunas ['b', 'c', 'f', 'l', 'm'], para as colunas numéricas vou realizar uma substituição utilizando a mediana, para isso vou utilizar o conjunto de validação.  

Colunas com até 70 valores únicos foram definidas como colunas do tipo categórico e colunas com mais foram definidas como numéricas.  


# Ajuste de Tipos

config_features = yaml.safe_load(open('../src/data/config/features.yml', 'r'))

fix_features = FixFeaturesType(config_features['fix_type_map'])

df_tmp = fix_features.transform(df)

# Ajuste de Missings

fix_missing = FixFeaturesMissing(config_features['fix_missing_numeric_features'],
                                 config_features['fix_missing_map'])

df_tmp = fix_missing.fit_transform(df_tmp)

df_tmp.info()
