==> notebooks/5-Feature-Engineering.ipynb <==
# Feature Engineering

import pandas as pd
from datetime import datetime

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 100)

# Data

df = pd.read_parquet('../data/interim/fraud_dataset_v2_valid.parquet.gzip')

df.head()

# Feature Engineering

def get_period_of_day(hour):
    if hour < 6:
        return 'dawn'
    if hour < 12:
        return 'morning'
    if hour < 18:
        return 'afternoon'
    if hour < 24:
        return 'evening'

class FeatureEngineering:
    """Class to create new features based on existing ones
    """
    def __init__(self, inference=True):
        self.inference = inference

    def fit(self, X, y=None):
        return self
    
    def create_train_features(self, X):
        X = X.reset_index(drop=True)
        X_tmp = X.reset_index(drop=True)

        X_tmp['ymd'] = X_tmp['fecha'].apply(lambda x: datetime.strftime(x, '%Y%m%d')).astype(int)

        new_columns = [col for col in X_tmp.columns if col not in X.columns]
        X_tmp = X.merge(X_tmp[new_columns], left_index=True, right_index=True, how='left')
        
        return X_tmp
    
    def create_payload_features(self, X):
        X = X.reset_index(drop=True)
        X_tmp = X.reset_index(drop=True)

        X_tmp['day_of_week'] = X_tmp['fecha'].dt.weekday

        X_tmp['hour_of_day'] = X_tmp['fecha'].dt.hour

        X_tmp['period_of_day'] = X_tmp['hour_of_day'].apply(get_period_of_day)

        new_columns = [col for col in X_tmp.columns if col not in X.columns]
        X_tmp = X.merge(X_tmp[new_columns], left_index=True, right_index=True, how='left')
        
        return X_tmp

    def transform(self, X):
        if not self.inference:
            X = self.create_train_features(X)
        return self.create_payload_features(X)
    
    def fit_transform(self, X, y=None):
        self.fit(X, y)
        return self.transform(X)

feature_eng = FeatureEngineering(inference=False)

df_tmp = feature_eng.fit_transform(df)

df_tmp.head()
