==> notebooks/7-Feature-Store.ipynb <==
# Feature Store

Após a análise exploratória tive algumas ideias para criação de novas variáveis, então nessa etapa serão criadas variáveis que não se encontram no dataset/payload e que podem ser construídas a partir de estruturas de feature stores.  

Nessa etapa será criado 1 *feature groups*:  
- Perfil Categoria Produto últimos 7 dias (as variáveis de fraude não serão tão realistas, visto que a maturação da fruade esta sendo desconsiderada)

A ideia desse feature group é trazer informações históricas dos produtos condensados.  

import pandas as pd
import numpy as np
from datetime import datetime


pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 100)

# Data

df = pd.concat([
    pd.read_parquet('../data/processed/fraud_dataset_v2_train.parquet.gzip'),
    pd.read_parquet('../data/processed/fraud_dataset_v2_valid.parquet.gzip'),
    pd.read_parquet('../data/processed/fraud_dataset_v2_test.parquet.gzip')
], ignore_index=True).reset_index(drop=True)

df.head(2)

# Features

## Perfil Estabelecimento

class FeatureStore:

    key = 'j'
    date_ref = 'ymd'
    window = 7
    delay = 1
    name = None
    path_to_save = None

    def __repr__(self):
        return f'I am {self.name}'

    def save(self):
        self.data.to_parquet(
            f'{self.path_to_save}/{self.name}.parquet.gzip',
            compression='gzip',
            index=False
        )

    def process(self, X):
        print('Hi! Implement this method!')

    def merge(self, X):
        return X.merge(
            self.data,
            how='left',
            on=[self.key, self.date_ref]
        )

class FeatureStoreCategoryProfile(FeatureStore):

    name = 'FeatureStoreCategoryProfile'
    path_to_save = '../data/feature_store'

    def process(self, df):
        # Fill dataset with all possible dates
        dt_init = str(df['ymd'].min())
        dt_end = str(df['ymd'].max())
        all_dates = pd.date_range(start=datetime.strptime(dt_init, '%Y%m%d'), end=datetime.strptime(dt_end, '%Y%m%d'))

        unique_clients = df['j'].unique()
        df_clients_with_all_dates = pd.MultiIndex.from_product([unique_clients, all_dates], names=['j', 'ymd']).to_frame(index=False)
        df_clients_with_all_dates['ymd'] = df_clients_with_all_dates['ymd'].dt.strftime('%Y%m%d').astype(int)

        df_processed = df_clients_with_all_dates.merge(df[['j', 'ymd', 'monto', 'fraude']], on=['j', 'ymd'], how='left')
        
        # Create new Features
        df_processed['monto_fraude'] = df_processed['monto'] * df_processed['fraude']

        # Aggregate
        self.data = (
            df_processed
            .groupby(['j', 'ymd'])
            .agg({
                'monto' : ['sum', 'count'],
                'monto_fraude' : 'sum',
                'fraude' : 'sum'
            })
            .sort_values(by=['j', 'ymd'], ascending=True)
            .reset_index()
            .groupby('j')
            .rolling(window=self.window, min_periods=1)
            .agg({
                ('ymd', '') : 'max',
                ('monto', 'sum') : 'sum',
                ('monto', 'count') : 'sum',
                ('monto_fraude', 'sum') : 'sum',
                ('fraude', 'sum') : 'sum'
            })
            .reset_index(level=0)
        )
        # Fix date ref, to not include current day
        self.data['ymd'] += self.delay
        self.data.columns = ['j', 'ymd', 'vl_cat_last7d', 'qty_cat_last7d', 'vl_fraud_cat_last7d', 'qty_fraud_cat_last7d']

        # New features to this feature store
        self.data['br_vl_cat_last7d'] = np.nan_to_num(self.data['vl_fraud_cat_last7d'] / self.data['vl_cat_last7d'], 
                                                       nan=0, posinf=1e9, neginf=-1e9)
        self.data['br_qty_cat_last7d'] = np.nan_to_num(self.data['qty_fraud_cat_last7d'] / self.data['qty_cat_last7d'], 
                                                       nan=0, posinf=1e9, neginf=-1e9)


fs = FeatureStoreCategoryProfile()

a = fs.process(df)

df[df['j'] == 'cat_0005972'].sort_values('ymd', ascending=True)[['monto', 'ymd']]

fs.data.head(20)

fs.save()

fs.merge(df)[['j', 'ymd', 'qty_cat_last7d', 'vl_cat_last7d']]

df[(df['j'] == 'cat_0cd53cb') & (df['ymd'] < 20200309)]['monto'].agg(['count', 'sum'])

Variável criada com sucesso e validada, foram implementadas duas classes para facilitar desenvolvimentos futuros, podendo herdar a classe FeatureStore e realizar apenas implementações específicas
