==> notebooks/8.Feature-Selection.ipynb <==
# Feature Selection

Mesmo com um problema com poucas variáveis vamos realizar um método de seleção de features.  
Vamos utilizar como forma de seleção o algoritmo RandomForest com os parametros a seguir:
- `n_estimators=100` : Quantidade razoaável de árvores  
- `criterion='gini'` : Eficiência computacional vs entropia  
- `max_depth=5` : Quantidade razoável de nós  
- `max_features='sqrt` : Oportunidade de interação de diferentes features  
- `class_weight='balanced'` : Problema desbalanceado, dando maior peso para classes minoritárias  
A escolha do algoritmo se deve ao fato de ser uma forma simples de avaliar a importância de cada uma das variáveis, avaliando um método de árvore (mesmo que será utilizado no algoritmo final) e selecioanando as variáveis, dando oportunidade para elas aparecerem em diferentes árvores interagindo com diferentes variáveis.  

No meio das variáveis serão colocadas 4 variáveis aleatórias, 2 categóricas (baixa cardinalidade) e 2 continuas.  
As variáveis selecionadas serão aquelas que se mantiverem acima da primeira aleatória no ranking de importância por `ganho de informação` ou que acumularem juntas 95% de importância (valem testes para definir esses valores).  

(método parecido com Boruta, porém computacionalmente mais rápido)

import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from feature_engine.encoding import OrdinalEncoder

# Data

df_train = pd.read_parquet('../data/enrich/fraud_dataset_v2_train.parquet.gzip')
df_valid = pd.read_parquet('../data/enrich/fraud_dataset_v2_valid.parquet.gzip')

for i in range(1,3):
    # replicability
    np.random.seed(i)
    df_train[f'___random_cat_{i}___'] = np.random.randint(0, i*25, size=df_train.shape[0])

    np.random.seed(i)
    df_train[f'___random_con_{i}___'] = np.random.uniform(0, 1000000, size=df_train.shape[0])

df_train.head()

# Feature Selection

hard_remove = ['c', 'p', 'q', 'fecha', 'i', 'score', 'fraude', 'ymd']

ord_enc = OrdinalEncoder(
    encoding_method='arbitrary',
    variables=df_train.drop(columns=hard_remove).select_dtypes(include=['object']).columns.tolist()
).fit(df_train)

rf = RandomForestClassifier(n_estimators=100, criterion='gini', max_depth=5,
                            max_features='sqrt', class_weight='balanced', random_state=777, n_jobs=-1)

rf.fit(ord_enc.transform(df_train).drop(columns=hard_remove), df_train['fraude'])

df_importance = pd.DataFrame({
    'feature': rf.feature_names_in_,
    'importance': rf.feature_importances_
}).sort_values(by='importance', ascending=False).reset_index(drop=True)
df_importance['importance_cumsum'] = df_importance['importance'].cumsum()
df_importance

Considerando o critério de acumular 95% de importância, temos 12 variáveis selecionadas:  
- o  
- f  
- n  
- l  
- a  
- m  
- s  
- br_qty_cat_last7d  
- d  
- br_vl_cat_last7d  
- qty_fraud_cat_last7d  
- h  
- vl_fraud_cat_last7d  

Interessante o valor da transação não ter entredo como variável, poderemos utilizar ele no futuro no sample weight do lightgbm para dar maior peso a transações de valor alto.
