==> notebooks/9-Encoding.ipynb <==
# Encoding

Nessa etapa vamos realizar o encoding das variáveis não numéricas, para conseguir passar os dados pelo modelo. Como vamos utilizar um modelo baseado em árvores, vamos utilizar um ordinal encoder, visto que o modelo consegue capturar não linearidade.  
Além dos encoders vamos criar todos os artefatos binários necessários para montar a pipeline do modelo.    


import pandas as pd
from feature_engine.encoding import OrdinalEncoder
import yaml
import pickle

import sys
sys.path.append('../')
from src.utils.transformers import Selector, FeatureEngineering, FixFeaturesMissing, FixFeaturesType

# Data

df = pd.read_parquet('../data/enrich/fraud_dataset_v2_train.parquet.gzip')

support_features = ['o','f','n','l','a','m','s','br_qty_cat_last7d',
                    'd','br_vl_cat_last7d','qty_fraud_cat_last7d','h',
                    'vl_fraud_cat_last7d']

config_features = yaml.safe_load(open('../src/data/config/features.yml', 'r'))

config_features

# Encoding

ord_enc = OrdinalEncoder(
    encoding_method='arbitrary',
    variables=df[support_features].select_dtypes(include=['object', 'string']).columns.tolist()
).fit(df)

ord_enc.transform(df).head(2)

selector = Selector(support_features).fit(df)
fix_type = FixFeaturesType(config_features['fix_type_map']).fit(df)
fix_missing = FixFeaturesMissing(config_features['fix_missing_numeric_features'], 
                                 config_features['fix_missing_map']).fit(df)
feature_engineering = FeatureEngineering().fit(df)

pickle.dump(ord_enc, open('../model/encoders/ordinal_encoder.pkl', 'wb'))
pickle.dump(selector, open('../model/encoders/selector.pkl', 'wb'))
pickle.dump(fix_type, open('../model/encoders/fix_type.pkl', 'wb'))
pickle.dump(fix_missing, open('../model/encoders/fix_missing.pkl', 'wb'))
pickle.dump(feature_engineering, open('../model/encoders/feature_engineering.pkl', 'wb'))
