==> notebooks/10-Tune.ipynb <==
# Tune

Trabalhando com um algoritmo de boosting temos diversos parametros que podem ser alterados, mudando a performance do modelo, nesta etapa utilizaremos a biblioteca `optuna` para tunagem dos hiperparametros. 

import pandas as pd
import lightgbm as lgbm
import optuna
import pickle

import sys
sys.path.append('../')

from src.utils.modeling import objective

# Data

df_train = pd.read_parquet('../data/encoded/fraud_dataset_v2_train.parquet.gzip')
df_valid = pd.read_parquet('../data/encoded/fraud_dataset_v2_valid.parquet.gzip')

selector = pickle.load(open('../model/encoders/selector.pkl', 'rb'))

# Tune

sampler = optuna.samplers.TPESampler(seed=777)
study = optuna.create_study(direction='maximize', sampler=sampler)

# Don't use n_jobs to keep replicability
study.optimize(lambda trial: objective(trial,
                                       selector.transform(df_train), 
                                       df_train['fraude'],
                                       [(selector.transform(df_valid), 
                                         df_valid['fraude'])]), n_trials=15, n_jobs=1)

print("Best parms:", study.best_params)

study.best_params
